FAT: moved find_file method from FileSlack to FAT implementation
