+++ fishy/fat/fat_filesystem/fat.py
@@ -295,3 +295,45 @@ class FAT:
                         dir_entry.start_cluster = int.from_bytes(dir_entry.firstCluster,
                                                                  byteorder='little')
                         yield (dir_entry, retlfn)
+
+    def find_file(self, filepath: str) -> DIR_ENTRY:
+        """
+        returns the directory entry for a given filepath
+        :param filepath: string, filepath to the file
+        :return: DIR_ENTRY of the requested file
+        """
+        # build up filepath as directory and
+        # reverse it so, that we can simple
+        # pop all filepath parts from it
+        path = filepath.split('/')
+        path = list(reversed(path))
+        # read root directory and append all entries
+        # as a tuple of (entry, lfn), that have a non
+        # empty lfn
+        # TODO: This excludes filesystems without long
+        #       filename extension. Should we support
+        #       them?
+        current_directory = []
+        for entry, lfn in self.get_root_dir_entries():
+            if lfn != "":
+                current_directory.append( (entry, lfn) )  # pylint: disable=bad-whitespace
+
+        while len(path) > 0:
+            fpart = path.pop()
+
+            # scan current directory for filename
+            filename_found = False
+            for entry, lfn in current_directory:
+                if lfn == fpart:
+                    filename_found = True
+                    break
+            if not filename_found:
+                raise Exception("File or directory '%s' not found" % fpart)
+
+            # if it is a subdirectory, enter it
+            if entry.attributes.subDirectory and len(path) > 0:
+                current_directory = []
+                for entry, lfn in self.get_dir_entries(entry.start_cluster):
+                    if lfn != "":
+                        current_directory.append( (entry, lfn) )  # pylint: disable=bad-whitespace
+        return entry
